Use two-level namespace for bundles on macOS
Use the two-level namespace for loadable modules on macOS instead of the
flat namespace. The flat namespace is a relic from Mac OS X 10.0 (2001).
The two-level namespace was introduced in Mac OS X 10.1 (2001) and the
ability to resolve undefined symbols at runtime in the two-level
namespace was added in Mac OS X 10.3 (2003). There's no reason to use
the flat namespace anymore today.