Make expt work with complex exponents
The following guard in the primitive definition for expt:
```
  if (negativep(y))
    return div2(MAKE_INT(1),
                my_expt(x, sub2(MAKE_INT(0), y)));
```
was preventing it from working with complex exponents, because
negativep will trigger an error for complexes. Adding `!COMPLEXP(y)`
fixes it.

However, there is another issue that we also fix with this PR: R7RS
says that

"The value of 0^z is 1 if (zero? z), 0 if (real-part z) is positive,
and an error otherwise.  Similarly for 0.0^z, with inexact results."

But the current code would not do the "zero when (real-part z) is
positive", so we also fix this in `my_expt`.

Then, we add some tests. And adjust one test from SRFI 70:

```
(test "expr.4" 1    (expt (sqrt 3) 0))
```

This won't work with R7RS, because `(sqrt 3)` is inexact, and then
`expt` is *required* to return an inexact number. So we do this instead:

```
(test "expr.4" 1.0    (expt (sqrt 3) 0)) ; because sqrt(3) is inexact
(test "expr.5" 1      (expt (sqrt 4) 0)) ; because sqrt(4) is exact
```

(And actually, as far as I can see, SRFI 70 doesn't really require that
the result from `(expt (sqrt 3) 0)` be exact).