Don't use test/error for compile errors
STklos does not have a method for catching compile-time errors
like syntax errors or malformed lets etc.

The previous test/error implementation was just ignoring the
code to be tested, so it made no difference. Now we use
`with-handler`, but it can't catch compile-time problems.

We have included a comment starting with 'TEST/ERROR' in all
the commented-out tests, so it's easier to restore them in
the future, if some error-catching mechanism is created for
compile-time problems.