Handle exports and imports properly in SRFI-94
SRFI 94 used to export `srfi-94:quotient`, and a few others.
This is not how it should be done: it should export plain
`quotient`, etc. The library/module system will take care
of everything. So now:

* SRFI-94 exports exactly what the SRFI spec says
* Its tests now run inside a module, so as to not change the
  core bindings that will be used in subsequent tests
* We do not define `square` in the tests (it's in R7RS)

This also fixes a failing test that happened because we
used the core versions of quotient & friends, and was
uncovered by the new test/error implementation.