Use if instead of do for multicommand macros
It is common practice to use

```
 #define SOME_MACRO do{ \
  ... \
  ... \
}while(0)
```

so that SOME_MACRO can be used inside `if`s and other C commands.

However, this can lead to problems:

```
 #define NEXT continue
...
...
for(;;) {
...
 #define LOCK_AND_RESTART                 do{\
  if (!have_global_lock) {                   \
    ...                                      \
    NEXT;                                    \
  }                                          \
}while(0)
...
} /* for(;;) */
```

The `NEXT` expands into `continue`, which will break out of the `do`,
and not continue the `for` loop as intended.

So we insteda use `if(1){ ... }else{}` for all multi-command macros.