Don't always convert reals to complexes
Because it causes problems like

```
(*  2+inf.0i  2)  =>  -nan.0+inf.0i
```

This should be 4+inf.0i, because the second argument has no imaginary
part, and the effect of the multiplication would be to just double the
real and imag parts of 2+inf.0i. However, since the number 2 is
promoted to complex, it becomes 2+0i, and the multiplication of 0 by
infinity results in NaN.

This is also the origin of this other problem:

```
(* 1 +inf.0+inf.0i) => -nan.0-nan.0i
```

And as a consequence:

```
(square +inf.0+1i) => +inf.0+inf.0i
(expt +inf.0+1i 2) => -nan.0-nan.0i
```

This is because (expt +inf.0+1i 2) calls exact_exponent_expt, which at
the end will multiply (* 1 +inf.0+inf.0i).

We add one more case in convert, with one more possible result, that
instead of a tc_ type, would mean "I did not convert, because one of
them is a complex and I'd add a zero that would be multiplied by a NaN
or infinity". For no particular reason, we have chosen tc_not_boxed
(could be any other non-numerical type).

With this patch, we have

```
(/ 2 +inf.0+0.0i) => +nan-0.0i
```

which is the same result given by Chez, Sagittarius and Chibi.