Complex -> integer when imag-part is exact 0
Currently, STklos does this:
```
(integer? (exact 65+0.0i)) => #f
```

Internally, a complex number is created with an exact zero
imaginary part... But since the zero is exact, it would be
strange to have this number as conplex: it doesn't satisfy
the `integer?` predicate, while 0+0i does:

```
(integer? 0+0i) => #t
```

An easy solution is to change `STk_inex2ex` to call `make_complex`
instead of `Cmake_complex`.