Adds a browseable manual from the REPL
Some Scheme implementations (Gauche and Gambit) will call an external
pager for the user to read the manual. This patch brings something
similar to STklos.

Added parameters:

* (default-browser) is a string with the full path of the browser to
  be used.

* (available-browsers) is a list of possible browsers (strings with
  executable names), that will be searched in the system PATH at boot
  time.

* (manual-path) is the full path to the STklos manual. At boot time,
  this is searched in
  (string-append (%library-prefix)
                 "/share/doc/stklos/"
                 (short-version)
                 "/html/stklos-ref.html")

Two procedures are added in `bonus.stk`:

`(open-in-browser url)`
Opens the URL given by the string |url| in the default browser, which
is determined by the string contained in the parameter
|default-browser|.

`(manual)`
Opens the STklos manual in a browser.