-nan.0 is now a synonym of +nan.0 as requested by R7RS
One (nasty) side effect is that evaluating -nan.0 returns +nan.0
(as Chibi, Guile, Petite, Gauche, Racket, ...)

   stklos> -nan.0
   +nan.0
   stklos> -nan.0-nan.0i
   +nan.0+nan.0i

The interest is that it permits to be more independant of the
underlying libc. For instance, before this patch (cos -inf.0) returns
-nan.0 (which is the valued returned by cos(-nan) ).
Manual says: "If x is positive infinity or negative infinity,
a domain error occurs, and a NaN is returned". But the sign of this
NaN is not stated. Anyway, other Schemes return +nan.0 in this case.