Adaptations for compiling STklos with tcc
It's damn fast (at compile time) but not very inefficient at runtime. Anyway,
it is a good thing for dev.

For now, tests on circular structures don't work (due to the produced code
which is not as compact as gcc one). Hope that correcting it will also
correct a long standing issue on macOS M1 architecture...