Modified the way the maximum recursive calls for equal? is built
tcc code is not compact and the actual divisor were not
adequate. Finally, we compute a value and fix MAX_EQUAL_CALLS as a
real max. Its values is now 30000, whereas we used a computed value of
more than 80000 before on x86_64 without problem. Maybe, it could be a
parameter object, or we could have a special case for tcc (but it
seems hat we have also a problem on Mac M1 or M2 -- not tested)