Reimplemented the do-repl-command function
Instead of having a fixed set of REL command, we can now add a new command
with repl-add-command.
For instance,

    (repl-add-command '(shell !)
                      "Run a shell command"
                      (lambda () (system (read-line))))

permits to add the ,! and ,shell REPL command.

The main interest is that a user can define its own commands in the `stklosrc`
file, if needed. This is a preliminary to implement some of the new commands
proposed by @jpellegrini in PR #599 (and perhaps not all of them, since this
cans be extended from outside of the REPL module).