Added tests to exhibit aliasing problems we have in STklos
Suppose we define
   (define-module A (export a) (define a 100))
and import the A module. Then, the "a" variable is read only as required by
R7RS. However, we can also redefine a in the importer module with
   (define a 1)
This will unset the RO bit on 'a', but the a in importer module is STILL
ALIASED to the "a" in module "A" (and changes in the importer change the value
in the A module).