Fixed aliasing bug.
All the global variables are in a unique pool. Module hash tables
contain the index of the variable in the pool (not a pointer since the
poll may be reallocated and, possibly, displaced).

Current implementation is minimal and is just here for
bootstraping (it was a painful job that took a looong time).

- Code must be simplified
- Optimizations to access global variables  must be added back
  (and adapted)
- Current implementation uses less memory and should be more efficient
  (hard to test now, since optimizations are deactivated).