Suppressed parameters (internal, imported, exported) from apropos
The problem is that result is not always accurate. For instance, if a symbol
is imported but redefined in a module, it will not be found by the apropos
command. This could be avoided, since it is possible to know if a symbol is an
alias or a locally defined variable. However it depends heavily on the current
implementation of global variables which will probably change in the future.

There is a place for another function, which tries to find information in
modules in an easier way than exploiting directly module introspection
functions.