Fix: be more strict with address-ref
GC_base returns NULL if this address is not within a region allocated by the
GC. But it also returns a non null value when we are inside an allocated
region (it returns the start of the allocated region). Consequently, we
consider that an address is valid if it is equal to the value returned by
GC_base.