Take care of infinites in or STk_mul2 or STk_div2
We use here some code from PR #587

However, we continue to always convert reals and integers to
complexes. This is different of the original code of the PR, which
modifies the convert function (and add it a third parameter).

Pros: The code is simpler, imo, and convert takes only 2 parameters as before.
Cons: We make sometime a useless conversion (but it the cases when we make
those conversions should be rare).