Protect the evaluation during constant folding
Otherwise, the compiler will produce an error at compile time.
==> If the evaluation produces an error, we stop to try folding.

Another option could be to produce a warning at compile time, but this is not
what is expected when we are running tests. For instance, in test-number we
have `(test/error "0^-a+bi" (expt 0 -2+1i))` and we expect to have a runtime
error. Building such test without warning would imply to use intermediary
variables and would be more complicate.