Suppressed code for generating IN-APPLY instruction
There was some code for producing IN-APPLY instruction in the code, since a
looong time. In fact, producing this instruction doesn't reduce call times
(in général we finish by replacing a GREF_INVOKE by an IN-APPLY, and that's
all). Note that we still need to produce a PREPARE_CALL, since the call to
apply can be a closure applied in tail position.