Rewrite time macro (do not use let-values anymore)
`let-values` expansion produces lambda expression which have an incidence on
the result. For instance, `(time #t)` says that we made 8
allocations (totalizing 296 bytes).

New version is still incorrect: we have
    stklos> (time #t)
    Elapsed time: 0.00999999999999801 ms
    Allocations: 16 bytes in 1 allocation calls
    #t

The 16 bytes correspond to the result of `(clock)` which is a boxed double.
We need a clock function which returns a boxed double.