Fix (atan 0.0 0) => +pi/2
And also (atan -0.0 0) => -pi/2

This is explicitly given as a corner case in R7RS (in the table in the
section for trigonometric procedures).

We also make (angle 0) trigger an error, since this would have to be
exactly the same as (atan (/ 0 0)), which can't be computed.

Interestingly, only Chez does trigger the same error (for "(angle 0)"),
but it is conceptually the right thing to do.

Tests included.