For errors ~S and ~W are equivalent (they use write*)
This fix the probleme signaled by @jpellegrini in PR #608.

Since, most of the time, we cannot assume that the culprit of an error is not
a circular object, we use `write*`, even when the format is a `~S` or a `~s`.
Changing the `~S` by `~W` was also possible, but it would have implied to
modify nearly all the source files.