Add `--without-keep-formals` option to configure
This adds a new option ot the configure script.

If the user passes `--without-keep-formals` to configure, then the
compiler `keep-formals` parameter will be true when compiling closures
in core STklos and all provided libraries.

However, this does not affect primitive procedures, because they are C
functions, and having the C compiler include the code to add the
formal parameters is, although possible... Tricky, at least.

How it works:

We have three autoconf variables:

```
    STR_KEEP_FORMALS="no"    # for output at the end of this script
    KEEP_FORMALS=0           # for tmpcomp (and in the future, C files)
    SCM_BOOT_KEEP_FORMALS="" # for the Makefile compiling the boot image
```

The tmpcomp file needs to be called in the Makefiles with the shell variable
```
KEEP_FORMALS set:
KEEP_FORMALS=@KEEP_FORMALS@ tmpcomp ...
```

For the boot image, the variable SCM_BOOT_KEEP_FORMALS is directly used.