Save the caller of an inlined primitive in a run_vm local variable
When we have
   (define (foo) (when (+ 1 1) (error "XXX")))

The error message was "+: XXX", because we changed the environment frame to
"+", to signal error if "+" parameters are incorrect. Otherwise, (+ 1 #f)
would yield the error "foo: '#f' is a bad number which is misleading.

Now, we restore the original caller after a successful return of the inlined
primitive.

NOTE: The restoration was not needed before we modified %build-error-location.