Adding a LET rewriter.
Basically, nested LETs and LET*s are joined into a LET*.

This is because going through several ENTER-LET instructions is
slower than a single ENTER-LET-STAR.

When do we *not* do this?

1. When the bindings of one LET shadows the bindings of the other

(LET ((a 1))
  (LET ((a 2)) body1) body2)

We obviously cannot turn this into
(LET* ((a 2)
       (a 2))
  body1
  body2)

2. When one individual binding of one of the LETs defines a variable,
but also uses its name in subsequent definitions:

(LET ((a 1)
      (x (1+ a))) ...)  ;; a is NOT the one defined above!

We cannot turn this into

(LET* ((a 1)
       (x (1+ a))) ...) ;; Oops, we now refer the the 'a' defined above!

3. One of the LETs is a named LET.

4. Finally: the expanion of LETREC is never optimized. This is because
when we do so we break test 1.2 of the "R5RS Pitfalls" suite.

How is it done?

- Item (4) above is attained by changing 'let' into '%let' in the
  expansion of LETREC (compiler.stk)
- The rest is done by a handful of small procedures in
  'comprewrite.stk', along with a new source rewriter for LET.