Simplify error checking
In the compiler, instead of

```
  (if some-condition
      ( X ...
          ...
          ... )
      (compiler-error Y))
```

we do

```
  (unless some-condition (compiler-error Y))
  ( X ...
      ...
      ... )
```

Or `when`, when it makes sense

This makes error checking isolated in the place where it
programatically happens, and also reduces source code size a little
bit, makeing it clearer.

There are already uses of the macros `when` and `unless` in the
compiler, so we're not introducing a dependency on them.