Optimize arithmetic for inexact real numbers
When we do "a + b" and one of the arguments is not real, it may
be converted to real by the 'convert' function (in number.c).

However,  `convert` will sometimes allocate a new cell for the
real number. But sometimes it is not necessary.

Then, STk_add2 will again allocate a new cell, in

```
o1 = double2real(REAL_VAL(o1) + REAL_VAL(o2));
```

We avoid the new allocation by checking first.

We do the same for all four operations (+, -, *, /).

Some benchmarks:

BEFORE:

```
stklos> (let ((x 1/2) (a 1.1)) (time (repeat 50_000_000 (set! a (+ 1.1 x )))) a)
;    Elapsed time: 9830.525 ms
;    Allocations: 1600000000 bytes in 100000000 allocation calls (GC: 1187)
1.6
stklos> (let ((x 1/2) (a 1.1)) (time (repeat 50_000_000 (set! a (* 1.1 x )))) a)
;    Elapsed time: 9446.878 ms
;    Allocations: 1600000000 bytes in 100000000 allocation calls (GC: 1184)
0.55
stklos> (let ((x 1/2) (a 1.1)) (time (repeat 50_000_000 (set! a (- 1.1 x )))) a)
;    Elapsed time: 9440.226 ms
;    Allocations: 1600000000 bytes in 100000000 allocation calls (GC: 1184)
0.6
stklos> (let ((x 1/2) (a 1.1)) (time (repeat 50_000_000 (set! a (/ 1.1 x )))) a)
;    Elapsed time: 9519.47 ms
;    Allocations: 1600000000 bytes in 100000000 allocation calls (GC: 1184)
2.2
```

AFTER:

```
stklos> (let ((x 1/2) (a 1.1)) (time (repeat 50_000_000 (set! a (+ 1.1 x )))) a)
;    Elapsed time: 5664.624 ms
;    Allocations: 800000000 bytes in 50000000 allocation calls (GC: 593)
1.6
stklos> (let ((x 1/2) (a 1.1)) (time (repeat 50_000_000 (set! a (* 1.1 x )))) a)
;    Elapsed time: 5792.377 ms
;    Allocations: 800000000 bytes in 50000000 allocation calls (GC: 592)
0.55
stklos> (let ((x 1/2) (a 1.1)) (time (repeat 50_000_000 (set! a (- 1.1 x )))) a)
;    Elapsed time: 5564.795 ms
;    Allocations: 800000000 bytes in 50000000 allocation calls (GC: 592)
0.6
stklos> (let ((x 1/2) (a 1.1)) (time (repeat 50_000_000 (set! a (/ 1.1 x )))) a)
;    Elapsed time: 5978.568 ms
;    Allocations: 800000000 bytes in 50000000 allocation calls (GC: 591)
2.2
```

So, basically we have somethiing like half the amount of allocations
and GC steps, and also something close to twice the speed, when we
operate on a real with some other type (different from real, and also
different from complex).

When we operate on two reals, there is no difference in timing.

Passes all tests.