Going back on commit a9dc556110bd02b2e1e078dd1490a0a051049b1e
This avoids an allocation when we have a mmultiplication by 0.0 or -0.0.
The problem here is that the tests are complicated. We have to test

   if (REAL_ZEROP(o1) && isnormal(REAL_VAL(o1))) return o2;
   if (REAL_ZEROP(o2) && isnormal(REAL_VAL(o2))) return o1;

for each addition to avoid the allocation of a real. Doing the addition and
allocate a number should be more efficient in general case (even if in some
case where we do an unnecessary allocation).