Add ansi-color-specifier?
So the user can check if an object is one of the symbols that specify
ANSI colors. This can be useful when writing code that takes those as
argument:

```
(define-method ascii-draw ((s <shape>) color)
 ;; we should check if color is a valid color here
 ...
 )
```

And... Since we're at it, why not add a method to obtain the full list
also? :)

```
(ansi-color-specifier-list)
  => (normal bold no-bold italic no-italic underline no-underline blink
      no-blink reverse no-reverse black bg-black red bg-red green
      bg-green yellow bg-yellow blue bg-blue magenta bg-magenta cyan
      bg-cyan white bg-white)
```

In order to do this, we include the `set!`s of all lambdas that define
procedures for ANSI color stuff under the lambda that defines the
color alist.