Changed the way default case sensitivity is set
It is no more possible to boot the system in case isnsensitive mode (since a
long time). Now, we always boot in case insensitive mode and swith the mode
just before being interactive. That means that image are always case
insensitive.

NOTE: read-case-sensitve change now the folding property of the current input
port (as if we had read a #fold-case of #!no-fold-case)