Don't use $@ when calling make recursively
This seems to break cross-copilation.

The program `tmpcomp` uses the environment variable STKLOS_BINARY
to tell which STklos executable it can use. However, when we use
this rule:

```
../streams/derived.ostk: ../streams/primitive.ostk:
	(cd ../streams && $(MAKE) $@)
```

that varible is not passed recursively to the child make, and
the process fails during cross-compilation.

If we do this:

```
../streams/derived.ostk:
	(cd ../streams && $(MAKE) derived.ostk)

../streams/primitive.ostk:
	(cd ../streams && $(MAKE) primitive.ostk)
```

it (magically?) works.