New VM instruction "SYMBOL-VALUE"
This instruction is accessible through the the special compiler form
    (%%symbol-value 'a 'STklos)
which will produce the following code
    000:  CONSTANT-PUSH        0
    002:  CONSTANT             1
    004:  SYMBOL-VALUE
    005:

    Constants:
    0: a
    1: STklos

This compiler form is used in `define-syntax` expansion which uses the
unexportded `find-clause` function defined in MBE module. In STklos modules,
this symbol is available through the symbol %find-macro-clause. However, in
pure R7RS libraries, this symbol is not available (as well as the find-module
function). The hack here consists to use %%symbol-value in the expandion of
define-syntax since it is always available even if not exported by any module.

NOTE: This special form can probaly be useful in other low level places where
we need to have an access to unexported symbols. Note that the current version
of %%symbol-value does not support the optional third parameter of
symbol-value.

This should correct the issue #705 signaled by @Retropikzel