Make eqv? and symbol=? agree.
Currently, if one of the symbols is uninterned, `eqv?` may consider
both to be the same. With this patch, `eqv?` won't deal with
uninterned symbols, rteating them always as "different from whatever
else".

This is the same behavior as Chez, Chicken, Gambit, Gauche, MIT.
(Bigloo, Kawa, Chibi do not have string->uninterned-symbol).

The change is:

1. In `eqv?`, check for hard equality (with `==`) so if one of the
   symbols is uninterned, the result is `#f`
2. Since we're at it, rewrite `symbol=?` so it will not accept
   non-symbol arguments (so its behavior will be the same as
   `string=?`

So now:
(eqv? (string->uninterned-symbol "x")
      (string->uninterned-symbol "x"))       => #f

This should fix #709